+++ frontend/wailsjs/go/main/App.d.ts
@@ -21,6 +21,8 @@ export function IsProtectionEnabled():Promise<boolean>;
 
 export function MinimizeToTray():Promise<void>;
 
+export function ResetAllowlist():Promise<void>;
+
 export function SaveConfig(arg1:config.Config):Promise<void>;
 
 export function SetProtectionEnabled(arg1:boolean):Promise<void>;